requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.2
idna==3.3
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.1
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
torch==1.12.1+cpu
torchaudio==0.12.1+cpu
torchvision==0.13.1+cpu
typing_extensions==4.3.0
urllib3==1.26.11

opencv-python~=4.7.0.68